Added "purge" command to uninstall script.

+++ uninstall.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # If something fails - exit
-set -e
+
 # Check if root
 if [[ whoami -ne "root" ]]; then
 	echo "Must be root to run this installer."
@@ -15,8 +15,11 @@ rm /usr/bin/bashshot.sh
 rm /usr/bin/bashshot_cleaner.sh
 rm /usr/bin/array
 
-# Removes configuration
-rm -r /etc/bashshot
+if [[ $1 == purge ]]
+then
+	# Removes configuration
+	rm -r /etc/bashshot
+fi
 
 # Removes file for log rotation
 rm /etc/logrotate.d/bashshot
